Delete will be changed. No Whitelabel Error Page

+++ src/main/java/com/YassineGroupSpringBoot/AppController/ApplicationController.java
@@ -36,58 +36,50 @@ public class ApplicationController {
     }
 
     @GetMapping("/show-users")
-    public String showAllUsers(HttpServletRequest request){
-        request.setAttribute("users",userService.showAllUsers());
+    public String showAllUsers(HttpServletRequest request) {
+        request.setAttribute("users", userService.showAllUsers());
         request.setAttribute("mode", "ALL_USERS");
         return "welcomepage";
     }
 
     @RequestMapping("/delete-user")
     public String deleteUser(@RequestParam int id, HttpServletRequest request) {
-        userService.deleteMyUser(id);
-        request.setAttribute("users", userService.showAllUsers());
-        request.setAttribute("mode", "ALL_USERS");
-        return "welcomepage";
+        if (userService.existUser(id)) {
+            userService.deleteMyUser(id);
+            request.setAttribute("users", userService.showAllUsers());
+            request.setAttribute("mode", "ALL_USERS");
+            return "welcomepage";
+        } else {
+            request.setAttribute("error", "Invalid Username or Password");
+            request.setAttribute("mode", "ALL_USERS");
+            return "welcomepage";
+        }
     }
 
     @RequestMapping("/edit-user")
-    public String editUser(@RequestParam int id, HttpServletRequest request){
+    public String editUser(@RequestParam int id, HttpServletRequest request) {
         request.setAttribute("user", userService.editUser(id));
-        request.setAttribute("mode","MODE_UPDATE");
+        request.setAttribute("mode", "MODE_UPDATE");
         return "welcomepage";
     }
 
     @RequestMapping("/login")
-    public String login(HttpServletRequest request){
-        request.setAttribute("mode","MODE_LOGIN");
+    public String login(HttpServletRequest request) {
+        request.setAttribute("mode", "MODE_LOGIN");
         return "welcomepage";
     }
 
     @RequestMapping("/login-user")
-    public String loginUser(@ModelAttribute User user, HttpServletRequest request){
-        if(userService.findByUsernameAndPassword(user.getUsername(),user.getPassword())!=null){
-            return"homepage";
-        }else{
-            request.setAttribute("error","Invalid Username or Password");
-            request.setAttribute("mode","MODE_LOGIN");
+    public String loginUser(@ModelAttribute User user, HttpServletRequest request) {
+        if (userService.findByUsernameAndPassword(user.getUsername(), user.getPassword()) != null) {
+            return "homepage";
+        } else {
+            request.setAttribute("error", "Invalid Username or Password");
+            request.setAttribute("mode", "MODE_LOGIN");
             return "welcomepage";
         }
 
     }
+
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
